#211 serialize Planets

+++ OctoAwesome/OctoAwesome.Basics/DebugMapGenerator.cs
@@ -31,7 +31,7 @@ namespace OctoAwesome.Basics
 
             for (int layer = 0; layer < planet.Size.Z; layer++)
                 result[layer] = new Chunk(new Index3(index.X, index.Y, layer), planet.Id);
-
+            
             int part = (planet.Size.Z * Chunk.CHUNKSIZE_Z) / 4;
 
             for (int y = 0; y < Chunk.CHUNKSIZE_Y; y++)
@@ -46,14 +46,13 @@ namespace OctoAwesome.Basics
                     {
                         if (z < (int)(height + part))
                         {
-                            int block = z % (Chunk.CHUNKSIZE_Z);
-                            int layer = (int)(z / Chunk.CHUNKSIZE_Z);
+                            int block = z % Chunk.CHUNKSIZE_Z;
+                            int layer = z / Chunk.CHUNKSIZE_Z;
                             result[layer].SetBlock(x, y, block, sandIndex);
                         }
                     }
                 }
             }
-
             column.CalculateHeights();
             return column;
         }
